move deploystorage portions to pkg/installer
