+++ pkg/installer/generateconfig.go
@@ -1,4 +1,4 @@
-package cluster
+package installer
 
 // Copyright (c) Microsoft Corporation.
 // Licensed under the Apache License 2.0.
